requirements.txt
playwright>=1.42.0
httpx>=0.27.0
openai>=1.14.0
pydantic>=2.6.0
torch>=2.2.0
torch-geometric>=2.5.0
# torch-scatter and torch-sparse are optional for PyG >=2.5
# Install separately after torch if needed: pip install torch-scatter torch-sparse
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
streamlit>=1.32.0
pandas>=2.2.0
pyarrow>=15.0.0
networkx>=3.2.1
umap-learn>=0.5.6
matplotlib>=3.8.3
pdfminer.six>=20221105
aiohttp>=3.9.3
tqdm>=4.66.2
python-dotenv>=1.0.1
sentence-transformers>=2.5.0
plotly>=5.19.0
